add check file exist write level

+++ core/tools/file_manager.rb
@@ -5,8 +5,10 @@ module File_manager
         return File.exist?(path_file) ? JSON.parse(File.read(path_file)) : nil
     end
     def self.write_level_file path_file, json
-        File.open(path_file,"w") do |file|
-            file.write(json.to_json)
+        if File.exist?(path_file)
+            File.open(path_file,"w") do |file|
+                file.write(json.to_json)
+            end
         end
     end
 end
